moved game form to be its own component
